LPS-111628 Remove generic DataLayoutBuilderSidebar and create FieldsSidebar component to render specific settigns for Fields Sidebar
