+++ modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/data_layout_builder/js/plugins/fields-sidebar/components/FieldsSidebar.es.js
@@ -12,10 +12,48 @@
  * details.
  */
 
-import React from 'react';
+import ClayForm from '@clayui/form';
+import React, {useContext, useState} from 'react';
 
-import DataLayoutBuilderSidebar from '../../../data-layout-builder/DataLayoutBuilderSidebar.es';
+import AppContext from '../../../AppContext.es';
+import Sidebar from '../../../components/sidebar/Sidebar.es';
+import FieldsSidebarBody from './FieldsSidebarBody.es';
+import FieldsSidebarSettingsBody from './FieldsSidebarSettingsBody.es';
+import FieldsSidebarSettingsHeader from './FieldsSidebarSettingsHeader.es';
 
-export default function() {
-	return <DataLayoutBuilderSidebar />;
+export default function({title}) {
+	const [{focusedCustomObjectField, focusedField}] = useContext(AppContext);
+	const [keywords, setKeywords] = useState('');
+
+	const hasFocusedField = Object.keys(focusedField).length > 0;
+	const hasFocusedCustomObjectField =
+		Object.keys(focusedCustomObjectField).length > 0;
+
+	const displaySettings = hasFocusedCustomObjectField || hasFocusedField;
+
+	return (
+		<Sidebar>
+			<Sidebar.Header>
+				<Sidebar.Title title={title} />
+
+				{displaySettings ? (
+					<FieldsSidebarSettingsHeader />
+				) : (
+					<ClayForm onSubmit={event => event.preventDefault()}>
+						<Sidebar.SearchInput
+							onSearch={keywords => setKeywords(keywords)}
+						/>
+					</ClayForm>
+				)}
+			</Sidebar.Header>
+
+			<Sidebar.Body>
+				{displaySettings ? (
+					<FieldsSidebarSettingsBody />
+				) : (
+					<FieldsSidebarBody keywords={keywords} />
+				)}
+			</Sidebar.Body>
+		</Sidebar>
+	);
 }
