+++ modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/data_layout_builder/js/plugins/fields-sidebar/components/FieldsSidebarSettingsHeader.es.js
@@ -1,0 +1,87 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+import ClayButton, {ClayButtonWithIcon} from '@clayui/button';
+import ClayDropDown from '@clayui/drop-down';
+import ClayIcon from '@clayui/icon';
+import {PagesVisitor} from 'dynamic-data-mapping-form-renderer';
+import React, {useContext} from 'react';
+
+import AppContext from '../../../AppContext.es';
+import DataLayoutBuilderContext from '../../../data-layout-builder/DataLayoutBuilderContext.es';
+
+export default function() {
+	const [{fieldTypes, focusedCustomObjectField, focusedField}] = useContext(
+		AppContext
+	);
+	let {settingsContext} = focusedField;
+
+	if (focusedCustomObjectField.settingsContext) {
+		settingsContext = focusedCustomObjectField.settingsContext;
+	}
+
+	const visitor = new PagesVisitor(settingsContext.pages);
+	const typeField = visitor.findField(field => field.fieldName === 'type');
+
+	const fieldType = fieldTypes.find(({name}) => {
+		return name === typeField.value;
+	});
+
+	const [dataLayoutBuilder] = useContext(DataLayoutBuilderContext);
+	const handleFocusedFieldBlur = () => {
+		dataLayoutBuilder.dispatch('sidebarFieldBlurred');
+	};
+
+	if (!fieldType) {
+		return null;
+	}
+
+	return (
+		<div className="d-flex">
+			<ClayButtonWithIcon
+				className="mr-2"
+				displayType="secondary"
+				monospaced={false}
+				onClick={handleFocusedFieldBlur}
+				symbol="angle-left"
+			/>
+
+			<ClayDropDown
+				className="d-inline-flex flex-grow-1"
+				onActiveChange={() => {}}
+				trigger={
+					<ClayButton
+						className="d-inline-flex flex-grow-1"
+						disabled={true}
+						displayType="secondary"
+					>
+						<ClayIcon
+							className="mr-2 mt-1"
+							symbol={fieldType.icon}
+						/>
+
+						{fieldType.label}
+
+						<span className="d-inline-flex ml-auto navbar-breakpoint-down-d-none pt-2">
+							<ClayIcon
+								className="inline-item inline-item-after"
+								symbol="caret-bottom"
+							/>
+						</span>
+					</ClayButton>
+				}
+			></ClayDropDown>
+		</div>
+	);
+}
